chore: convert more files to ts
